pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-wiki-builder"
version = "1.0.0"
description = "AI-powered GitHub Wiki generator for automated documentation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Git Wiki Builder Team", email = "support@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup",
]
keywords = ["github", "wiki", "documentation", "ai", "automation"]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "gitpython>=3.1.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pathspec>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "bandit>=1.7.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
git-wiki-builder = "git_wiki_builder.cli:main"

[project.urls]
Homepage = "https://github.com/example/git-wiki-builder"
Repository = "https://github.com/example/git-wiki-builder"
Issues = "https://github.com/example/git-wiki-builder/issues"
Documentation = "https://github.com/example/git-wiki-builder/wiki"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
git_wiki_builder = ["templates/*", "prompts/*"]

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 79
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=src/git_wiki_builder",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
